+++ lua/plugins/status-line.lua
@@ -1,9 +1,0 @@
--- Status line plugins.
-
-return {
-  {
-    "nvim-lualine/lualine.nvim",
-    dependencies = { "nvim-web-devicons" },
-    config = true,
-  },
-}